refactor: simplify shared assistant shell and pane ownership
